# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.APIError:
    properties:
      details: {}
      message:
        type: string
    type: object
  handler.APISuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
  handler.AuthResponse:
    properties:
      expiresAt:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/handler.UserResponse'
    type: object
  handler.AuthSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/handler.AuthResponse'
      message:
        type: string
    type: object
...
  handler.CategoryItemSuccess:
    properties:
      data:
        $ref: '#/definitions/handler.CategoryResponse'
      message:
        type: string
    type: object
  handler.CategoryListSuccess:
    properties:
      data:
        items:
          $ref: '#/definitions/handler.CategoryResponse'
        type: array
      message:
        type: string
    type: object
  handler.CategoryRequest:
    properties:
      active:
        type: boolean
      colorHex:
        type: string
      icon:
        type: string
      name:
        type: string
      order:
        type: integer
      type:
        type: string
    type: object
  handler.CategoryResponse:
    properties:
      active:
        type: boolean
      colorHex:
        type: string
      createdAt:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      order:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.LogoutRequest:
    properties:
      allDevices:
        type: boolean
    type: object
...
  handler.RegisterRequest:
    properties:
      currency:
        type: string
      email:
        type: string
      language:
        type: string
      monthlyLimit:
        type: number
      name:
        type: string
      password:
        type: string
      theme:
        type: string
    type: object
...
  handler.UserConfigResponse:
    properties:
      currency:
        type: string
      language:
        type: string
      monthlyLimit:
        type: number
      notificationsEnabled:
        type: boolean
      theme:
        type: string
    type: object
  handler.UserProfileSuccess:
    properties:
      data:
        $ref: '#/definitions/handler.UserResponse'
      message:
        type: string
    type: object
  handler.UserResponse:
    properties:
      active:
        type: boolean
      config:
        $ref: '#/definitions/handler.UserConfigResponse'
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastLogin:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: contato@example.com
    name: Equipe Golang Finance
  description: API para gestão financeira pessoal com suporte a OCR de recibos.
  title: Golang Finance API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Valida credenciais e emite um novo token de sessão
      parameters:
      - description: Credenciais
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      summary: Autenticar usuário
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalida a sessão atual ou todas as sessões do usuário autenticado
      parameters:
      - description: Opções de encerramento
        in: body
        name: body
        schema:
          $ref: '#/definitions/handler.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APISuccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Encerrar sessão
      tags:
      - Auth
  /auth/me:
    get:
      description: Retorna os dados do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserProfileSuccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Perfil do usuário
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Cria um usuário, configura dados padrão e retorna a sessão autenticada
      parameters:
      - description: Dados de registro
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      summary: Registrar novo usuário
      tags:
      - Auth
  /categories:
    get:
      description: Retorna as categorias do usuário autenticado
      parameters:
      - description: Filtra por status das categorias. Use true para apenas ativas,
          false para apenas inativas e deixe em branco ou use all para todas.
        enum:
        - "true"
        - "false"
        - all
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CategoryListSuccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Listar categorias
      tags:
      - Categorias
    post:
      consumes:
      - application/json
      description: Cria uma categoria personalizada para o usuário
      parameters:
      - description: Dados da categoria
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CategoryItemSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Criar categoria
      tags:
      - Categorias
  /categories/{id}:
    delete:
      description: Marca uma categoria como inativa
      parameters:
      - description: Identificador da categoria
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APISuccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Desativar categoria
      tags:
      - Categorias
    put:
      consumes:
      - application/json
      description: Atualiza dados de uma categoria existente
      parameters:
      - description: Identificador da categoria
        in: path
        name: id
        required: true
        type: string
      - description: Campos para atualização
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CategoryItemSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Atualizar categoria
      tags:
      - Categorias
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"